validate not null input parameters on material controller

+++ src/main/java/com/enrique/reservatusalaback/controller/MaterialController.java
@@ -25,6 +25,9 @@ public class MaterialController {
 
     @PostMapping
     public ResponseEntity<?> add(@RequestBody Material material) {
+        if (material == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_MATERIAL, HttpStatus.BAD_REQUEST);
+        }
         return ResponseEntity.ok(materialService.add(material));
     }
 
@@ -35,6 +38,9 @@ public class MaterialController {
 
     @GetMapping("/findById")
     public ResponseEntity<?> findById(@RequestParam Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
         Material material = materialService.findById(id);
         if (material == null){
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
@@ -44,6 +50,9 @@ public class MaterialController {
 
     @PutMapping
     public ResponseEntity<?> update(@RequestBody Material material) {
+        if (material == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_MATERIAL, HttpStatus.BAD_REQUEST);
+        }
         Material updatedMaterial = materialService.update(material);
         if (updatedMaterial == null) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
@@ -53,6 +62,9 @@ public class MaterialController {
 
     @DeleteMapping
     public ResponseEntity<ResponseCode> deleteById(@RequestParam Long id) {
+        if (id == null) {
+            return new ResponseEntity<>(ResponseCode.NULL_ID, HttpStatus.BAD_REQUEST);
+        }
         if (materialService.deleteById(id) > 0) {
             return new ResponseEntity<>(ResponseCode.NOT_FOUND_ID, HttpStatus.NOT_FOUND);
         } else {
